+++ packages/cs-shared/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.20.6] - 2021-11-05
+### Fixed
  - fixed refresh of UTableEntity what based on Repository with subquery (by correctly skip handling of subquery on the client side)
 
 ## [5.20.5] - 2021-09-24